move vis to another file connect it to the backend
